Overhaul TextMate syntax highlighting (#26)
Continuation of #23. The current TextMate language specification is almost directly based on Swift, so this PR fixes most of the remaining syntax highlighting and namespace issues.

This PR also adds a minimal `.vscodeignore` file as encouraged by the `vsce package` command. 

Works as expected on Ubuntu (WSL) using the latest version of VSCode.