[VSC-7] Fix dependency analysis
Problem: We need to pass the filename to the import resolver, which
requires passing the filename to `parseMotokoWithDeps`. Also, we need to
notify files in the correct oder.

Solution: Parse the file name to `parseMotokoWithDeps`. Ensure that we
first write all files and only then we notify, otherwise we'll try to
analyse dependencies that weren't written yet (don't exist in virtual
filesystem).