[VSC-10] Support finding references of mod fields
Problem: We'd like to find references of module fields, e.g.,
`M1.M2./*...*/.Mn.f` where `M1` to `Mn` (n > 0) are modules and `f` is a
field.

Solution: Implement a handler for references that will try to find the
definition D of a variable under the cursor and for each AST in the
project, find all nodes that have the same definition as D. The search
is first narrowed down to nodes with the same name, and then to nodes
with the same definition, in order to avoid looking for definitions for
everything in the AST. Tests were added to confirm it's working as
intended.