[VSC-9] Allow completions with multiple lines
Problem: Completions current only work if the entirety of it is written
in a single line.

Solution: Rather than indexing the line at the cursor's line, we instead
match everything between the start of the document and the cursor's
line.

Also write some tests to check that it's working as intended.