[VSC-10] Support type search for fields
Problem: To look for `f` in `o.f`, we should support checking for `f` in
the type of `o`.

Solution: Enable combining sources in the typechecker and consider
`typeRep` from the compiler and look for matching fields when searchnig
for definition in a `DotE` expression. Add a lot of tests.