Add error recovery & module autocompletion (Serokell, Milestone 2) (#334)
* add tests for parseMotoko with error recovery

* enable error recovery in AST resolver

* [VSC-9] Export some navigation functions

* [VSC-9] Populate importable fields cache

* [VSC-9] Support module autocompletions

* [VSC-9] Fix kind for module names completions

* [VSC-9] Add test helper to wait for notification

* [VSC-9] Add completion test

* [VSC-9] Allow completions with multiple lines

Problem: Completions current only work if the entirety of it is written
in a single line.

Solution: Rather than indexing the line at the cursor's line, we instead
match everything between the start of the document and the cursor's
line.

Also write some tests to check that it's working as intended.

* fixup! [VSC-9] Support module autocompletions

* Update moc to 0.14.8

---------

Co-authored-by: Savely Krendelhoff <alicecandyom@proton.me>
Co-authored-by: Heitor Toledo Lassarote de Paula <heitor.toledo@serokell.io>