[VSC-22] Cache scores for scoreFn
Problem: In `findMostSpecificNodeForPosition`, we compute `scoreFn(n)`
and `scoreFn(node)` multiple times. However, under the condition that
`scoreFn` is pure/deterministic, we can cache it to avoid recomputing it
each time.

Solution: Assume `scoreFn` is pure/deterministic to be cached and if the
condition is met, save it. Now 6 calls to `scoreFn` became just 1.