Fix corner cases of go-to-definition & support references for object field (Serokell, Milestone-3) (#347)
* [VSC-10] Support finding references of mod fields

Problem: We'd like to find references of module fields, e.g.,
`M1.M2./*...*/.Mn.f` where `M1` to `Mn` (n > 0) are modules and `f` is a
field.

Solution: Implement a handler for references that will try to find the
definition D of a variable under the cursor and for each AST in the
project, find all nodes that have the same definition as D. The search
is first narrowed down to nodes with the same name, and then to nodes
with the same definition, in order to avoid looking for definitions for
everything in the AST. Tests were added to confirm it's working as
intended.

* [VSC-10] Support ID in AST

Problem: We exposed names in the AST as nodes. This is so we can get the
locations of these symbols.

Solution: Change the language server to take advantage of those new
nodes. This is also so existing tests will continue passing, as they
currently don't with these changes.

* [VSC-14] Add tests for go to definition

* [VSC-15] Fix file uri cache population

Problem: During resolve of the import information, uri
is analyzed. If uri points to `lib` file, `lib` suffix
was stripped from the uri. But after that regex expects this
`lib` suffix to be present in order to infer correct info.

Solution: Don't strip `lib` suffix.

* [VSC-14] Return correct cursor for type search

Problem: Cursor of the definition was set to the
type variable, instead of the type declaration.

Solution: Set it to the type declaration.

* [VSC-14,15,16] fix some corner-cases of go-to-definition

* comment out some WIP reference tests

* [VSC-10] Support type search for fields

Problem: To look for `f` in `o.f`, we should support checking for `f` in
the type of `o`.

Solution: Enable combining sources in the typechecker and consider
`typeRep` from the compiler and look for matching fields when searchnig
for definition in a `DotE` expression. Add a lot of tests.

* [VSC-10] Support multiple definitions

Problem: We can find multiple definitions from a type search.

Solution: Change `findDefinition` to return a list of definitions.
Rename to `findDefinitions`. Propagate changes while trying to not be
very intrusive to existing code.

* [VSC-10] Add a test for corner cases

Problem: We should test that with corner cases we get some reasonable
locations.

Solution: Add a test for a more complex file.

* [VSC-10] Reorder node priorities

Problem: Node priorities could be ordered from greatest to lower
priority.

Solution: Move ImportE to the end.

* [VSC-10] Include more definitions in references

Problem: It's possible that when we look for definitions for the
references in the AST, we might encounter more definitions we didn't see
before.

Solution: If the context asks to include declarations, then also include
the locations for the IDs of the new definitions.

* [VSC-10] Refactor test suites

Problem: We have two files for testing definitions, so we should join
into one. Moreover, we have a lot of shared things between the
definition and reference tests.

Solution: Join test cases from `src/server/definition.spec.ts` into
`src/server/test/definition.spec.ts`. Join Motoko reference test files
tests into the Motoko definitions test files. Extract common parts to
helpers. Move reference tests to `src/server/test` directory.

* [VSC-10] Handle review and latest moc

Problem: With the review on motoko we need to handle one test that isn't
valid anymore.

Solution: Comment the test. Also, address the review on making `docs`
constant again.

* [VSC-10] Remove DotE references workaround

Problem: During our work, this turned out to be redundant.

Solution: Remove the `DotE` -> `VarE` transformation. In `motoko`, we
also do not serialize the `typeRep` for `DotE` as a consequence.

* [VSC-22] Simplify return type for findDefinitions

Problem: `findDefinitions` returns `Definition | Definition[] |
undefined`, which is more complex than it needs to be.

Solution: Replace the return type to simply `Definition[]` and propagate
changes.

* [VSC-22] Cache scores for scoreFn

Problem: In `findMostSpecificNodeForPosition`, we compute `scoreFn(n)`
and `scoreFn(node)` multiple times. However, under the condition that
`scoreFn` is pure/deterministic, we can cache it to avoid recomputing it
each time.

Solution: Assume `scoreFn` is pure/deterministic to be cached and if the
condition is met, save it. Now 6 calls to `scoreFn` became just 1.

* Rename 'A.mo' (fix quirk with Git diff logic)

* Rename back to 'A.mo'

* Update 'motoko' dependency

* Rename 'a.mo' to 'simple.mo'

* Roll back changes to 'completion.spec.ts'

* [VSC-22] Increase test timeout

Problem: Tests on CI are failing, sounds like 10 s is not enough.

Solution: Increase timeouts to 60 s.

* [VSC-22] Fix timeouts in server spec

Problem: Looks like the timeout must be outside `beforeAll`. Hopefully
it should go from 5s to 60s now.

Solution: Move timeout setting to outside `beforeAll`. Also, remove a
timeout from one of the tests from 10s so it picks up the 60s timeout.

---------

Co-authored-by: Heitor Toledo Lassarote de Paula <heitor.toledo@serokell.io>
Co-authored-by: Savely Krendelhoff <alicecandyom@proton.me>
Co-authored-by: rvanasa <ryan.vandersmith@dfinity.org>