[VSC-34] Improve references
Problem: We don't always find references. For example, when setting a
field in an object, or when searching for an object field. Moreover, we
don't always remove definitions correctly when searching for references.

Solution: Delete definitions for newly-found references. Always search
the type representation, as we might be dealing with fields. Point the
cursor of an ExpField to the ID and not the the argument. Match more
node types when searching, like fields. Add tests to support these
changes.