[VSC-39] Implement rename
Problem: Now that prepare rename is implemented, we can implement
rename.

Solution: This request is simple: we find all references of whatever is
under the cursor, and create an edit for all of them. We add tests to
ensure this capability works as intended, by reusing both the
definitions and prepare rename tests.