LSP -- Improve Completions And Improve References -- Serokell Grant 2 Milestone 2 (#408)
* [VSC-33] Improve completion

Problems:
1. Completion list contains not relevant identifiers
2. Completion item details contains only module name
3. Completion items lack of documentation.

Solution:
1. Analyze code block containing the current cursor position to show
only relevant identifiers.
2. When item type is available show the item type as a detail.
3. Show documentation if it exists for the completion item.

* [VSC-34] Improve references

Problem: We don't always find references. For example, when setting a
field in an object, or when searching for an object field. Moreover, we
don't always remove definitions correctly when searching for references.

Solution: Delete definitions for newly-found references. Always search
the type representation, as we might be dealing with fields. Point the
cursor of an ExpField to the ID and not the the argument. Match more
node types when searching, like fields. Add tests to support these
changes.

* Bump motoko version

Problem: We need to use motoko 3.16.0 which contains relevant changes
for this milestone.

Solution: From from 3.15.1 to 3.16.0.

---------

Co-authored-by: Alexey Danilevsky <alexey.danilevsky@gmail.com>