[VSC-42] Add a test for local and nested modules
Problem: Now that we have better support for recovering the typed AST
for dot expressions in the compiler, we should also add a test for it in
the extension.

Solution: Add a test checking that we can complete from a local module
and another that we can complete from a nested module. While testing, I
found that we don't properly filter items based on their prefix, so I
added this case for completions from definitions.