[VSC-42] Do not filter items server side
Problem: According to the LSP specification, the server should not
filter items unless we send `isIncomplete` is `true`. However, this
would cause the client to recompute the list on every keystroke, and
since we rely on `findDefinitions`, which is slow, we don't want this.

Solution: Do not filter items based on a prefix. To check for the items,
order, and continuity in a subarray, we had to write a custom Jest
matcher. Set `isIncomplete` as `false` since we now send the complete
list.